Remove variant & telemetry support for .NET Feature Management schema (#421)
* reorg test suite

* remove variant and telemetry & reorg test suite

* update

* update constructor

* update appsettings.json of EvaluationDataToApplicationInsights example project

* update appsetting.json for all examples
