+++ examples/BlazorServerApp/appsettings.json
@@ -6,33 +6,43 @@
     }
   },
   "AllowedHosts": "*",
-  "FeatureManagement": {
-    "BrowserEnhancement": {
-      "EnabledFor": [
-        {
-          "Name": "Browser",
-          "Parameters": {
-            "AllowedBrowsers": [ "Edge" ]
-          }
-        }
-      ]
-    },
-    "Beta": {
-      "EnabledFor": [
-        {
-          "Name": "Targeting",
-          "Parameters": {
-            "Audience": {
-              "DefaultRolloutPercentage": 50,
-              "Exclusion": {
-                "Groups": [
-                  "Guests"
-                ]
+  "feature_management": {
+    "feature_flags": [
+      {
+        "id": "BrowserEnhancement",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "Browser",
+              "parameters": {
+                "AllowedBrowsers": [ "Edge" ]
               }
             }
-          }
+          ]
         }
-      ]
-    }
+      },
+      {
+        "id": "Beta",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "Targeting",
+              "parameters": {
+                "Audience": {
+                  "DefaultRolloutPercentage": 50,
+                  "Exclusion": {
+                    "Groups": [
+                      "Guests"
+                    ]
+                  }
+                }
+              }
+            }
+          ]
+        }
+      }
+    ]
   }
 }
