+++ examples/ConsoleApp/appsettings.json
@@ -1,14 +1,20 @@
 ﻿{
-  "FeatureManagement": {
-    "Beta": {
-      "EnabledFor": [
-        {
-          "Name": "AccountId",
-          "Parameters": {
-            "AllowedAccounts": [ "abcdefghijklmnopqrstuvwxyz" ]
-          }
+  "feature_management": {
+    "feature_flags": [
+      {
+        "id": "Beta",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "AccountId",
+              "parameters": {
+                "AllowedAccounts": [ "abcdefghijklmnopqrstuvwxyz" ]
+              }
+            }
+          ]
         }
-      ]
-    }
+      }
+    ]
   }
 }
