+++ examples/EvaluationDataToApplicationInsights/appsettings.json
@@ -1,62 +1,61 @@
 {
-    "Logging": {
-        "LogLevel": {
-            "Default": "Information",
-            "Microsoft.AspNetCore": "Warning"
-        }
-    },
-    "AllowedHosts": "*",
-    "ApplicationInsights": {
-        "ConnectionString": "<connection-string>"
-    },
-    "FeatureManagement": {
-        "ImageRating": {
-            "Telemetry": {
-                "Enabled": true,
-                "Metadata": {
-                    "Label": "A Label",
-                    "Etag": "An etag",
-                    "Tags.A": "Tag A value"
-                }
-            },
-            "Variants": [
-                {
-                    "Name": "BlackAndWhite",
-                    "ConfigurationValue": "https://cdn.pixabay.com/photo/2017/09/28/09/26/dinosaur-2794840_1280.png"
-                },
-                {
-                    "Name": "WithColor",
-                    "ConfigurationValue": "https://cdn.pixabay.com/photo/2012/04/02/15/18/stegosaurus-24752_1280.png"
-                },
-                {
-                    "Name": "HighQuality",
-                    "ConfigurationValue": "https://cdn.pixabay.com/photo/2019/02/01/14/24/landscape-3969074_1280.jpg"
-                }
-            ],
-            "Allocation": {
-                "Percentile": [
-                    {
-                        "Variant": "BlackAndWhite",
-                        "From": 0,
-                        "To": 33
-                    },
-                    {
-                        "Variant": "WithColor",
-                        "From": 33,
-                        "To": 66
-                    },
-                    {
-                        "Variant": "HighQuality",
-                        "From": 66,
-                        "To": 100
-                    }
-                ]
-            },
-            "EnabledFor": [
-                {
-                    "Name": "AlwaysOn"
-                }
-            ]
-        }
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "AllowedHosts": "*",
+  "ApplicationInsights": {
+    "ConnectionString": "<connection-string>"
+  },
+  "feature_management": {
+    "feature_flags": [
+      {
+        "id": "ImageRating",
+        "enabled": true,
+        "telemetry": {
+          "enabled": true,
+          "metadata": {
+            "Label": "A Label",
+            "Etag": "An etag",
+            "Tags.A": "Tag A value"
+          }
+        },
+        "variants": [
+          {
+            "name": "BlackAndWhite",
+            "configuration_value": "https://cdn.pixabay.com/photo/2017/09/28/09/26/dinosaur-2794840_1280.png"
+          },
+          {
+            "name": "WithColor",
+            "configuration_value": "https://cdn.pixabay.com/photo/2012/04/02/15/18/stegosaurus-24752_1280.png"
+          },
+          {
+            "name": "HighQuality",
+            "configuration_value": "https://cdn.pixabay.com/photo/2019/02/01/14/24/landscape-3969074_1280.jpg"
+          }
+        ],
+        "allocation": {
+          "percentile": [
+            {
+              "variant": "BlackAndWhite",
+              "from": 0,
+              "to": 33
+            },
+            {
+              "variant": "WithColor",
+              "from": 33,
+              "to": 66
+            },
+            {
+              "variant": "HighQuality",
+              "from": 66,
+              "to": 100
+            }
+          ]
+        }
+      }
+    ]
+  }
 }
