+++ examples/FeatureFlagDemo/appsettings.json
@@ -7,65 +7,85 @@
   "AllowedHosts": "*",
 
   // Define feature flags in config file
-  "FeatureManagement": {
-
-    "Home": true,
-    "Beta": {
-      "EnabledFor": [
-        {
-          "Name": "Targeting",
-          "Parameters": { // This json object maps to a strongly typed configuration class
-            "Audience": {
-              "Users": [
-                "Jeff",
-                "Alicia"
-              ],
-              "Groups": [
-                {
-                  "Name": "Ring0",
-                  "RolloutPercentage": 80
-                },
-                {
-                  "Name": "Ring1",
-                  "RolloutPercentage": 50
+  "feature_management": {
+    "feature_flags": [
+      {
+        "id": "Home",
+        "enabled": true
+      },
+      {
+        "id": "Beta",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "Microsoft.Targeting",
+              "parameters": {
+                "Audience": {
+                  "Users": [
+                    "Jeff",
+                    "Alicia"
+                  ],
+                  "Groups": [
+                    {
+                      "Name": "Ring0",
+                      "RolloutPercentage": 80
+                    },
+                    {
+                      "Name": "Ring1",
+                      "RolloutPercentage": 50
+                    }
+                  ],
+                  "DefaultRolloutPercentage": 20
                 }
-              ],
-              "DefaultRolloutPercentage": 20
+              }
             }
-          }
+          ]
         }
-      ]
-    },
-    "CustomViewData": {
-      "EnabledFor": [
-        {
-          "Name": "Browser",
-          "Parameters": {
-            "AllowedBrowsers": [ "Chrome", "Edge" ]
-          }
+      },
+      {
+        "id": "CustomViewData",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "Browser",
+              "parameters": {
+                "AllowedBrowsers": [ "Chrome", "Edge" ]
+              }
+            }
+          ]
         }
-      ]
-    },
-    "ContentEnhancement": {
-      "EnabledFor": [
-        {
-          "Name": "TimeWindow",
-          "Parameters": {
-            "Start": "Wed, 01 May 2019 13:59:59 GMT",
-            "End": "Mon, 01 July 2019 00:00:00 GMT"
-          }
+      },
+      {
+        "id": "ContentEnhancement",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "Microsoft.TimeWindow",
+              "parameters": {
+                "Start": "Wed, 01 May 2019 13:59:59 GMT",
+                "End": "Mon, 01 Jul 2019 00:00:00 GMT"
+              }
+            }
+          ]
         }
-      ]
-    },
-    "EnhancedPipeline": {
-      "EnabledFor": [
-        {
-          "Name": "Microsoft.Percentage",
-          "Parameters": {
-            "Value": 50
-          }
+      },
+      {
+        "id": "EnhancedPipeline",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "Microsoft.Percentage",
+              "parameters": {
+                "Value": 50
+              }
+            }
+          ]
         }
-      ]
-    }
+      }
+    ]
   }
 }
