+++ examples/RazorPages/appsettings.json
@@ -6,7 +6,12 @@
     }
   },
   "AllowedHosts": "*",
-  "FeatureManagement": {
-    "Home": true
+  "feature_management": {
+    "feature_flags": [
+      {
+        "id": "Home",
+        "enabled": true
+      }
+    ]
   }
 }
