+++ examples/TargetingConsoleApp/appsettings.json
@@ -1,39 +1,45 @@
 ﻿{
-  "FeatureManagement": {
-    "Beta": {
-      "EnabledFor": [
-        {
-          "Name": "Targeting",
-          "Parameters": {
-            "Audience": {
-              "Users": [
-                "Jeff",
-                "Anne"
-              ],
-              "Groups": [
-                {
-                  "Name": "Management",
-                  "RolloutPercentage": 100
-                },
-                {
-                  "Name": "TeamMembers",
-                  "RolloutPercentage": 45
+  "feature_management": {
+    "feature_flags": [
+      {
+        "id": "Beta",
+        "enabled": true,
+        "conditions": {
+          "client_filters": [
+            {
+              "name": "Microsoft.Targeting",
+              "parameters": {
+                "Audience": {
+                  "Users": [
+                    "Jeff",
+                    "Anne"
+                  ],
+                  "Groups": [
+                    {
+                      "Name": "Management",
+                      "RolloutPercentage": 100
+                    },
+                    {
+                      "Name": "TeamMembers",
+                      "RolloutPercentage": 45
+                    }
+                  ],
+                  "DefaultRolloutPercentage": 20,
+                  "Exclusion": {
+                    "Users": [
+                      "Anne",
+                      "Phil"
+                    ],
+                    "Groups": [
+                      "Contractor"
+                    ]
+                  }
                 }
-              ],
-              "DefaultRolloutPercentage": 20,
-              "Exclusion": {
-                "Users": [
-                  "Anne",
-                  "Phil"
-                ],
-                "Groups": [
-                  "Contractor"
-                ]
               }
             }
-          }
+          ]
         }
-      ]
-    }
+      }
+    ]
   }
 }
