+++ src/Microsoft.FeatureManagement/DotnetFeatureManagementFields.cs
@@ -1,0 +1,19 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+//
+
+namespace Microsoft.FeatureManagement
+{
+    internal static class DotnetFeatureManagementFields
+    {
+        public const string RequirementType = "RequirementType";
+
+        // Feature filters keywords
+        public const string FeatureFiltersSectionName = "EnabledFor";
+        public const string FeatureFilterConfigurationParameters = "Parameters";
+
+        // Other keywords
+        public const string NameKeyword = "Name";
+        public const string FeatureManagementSectionName = "FeatureManagement";
+    }
+}
