+++ tests/Tests.FeatureManagement/Features.cs
@@ -9,16 +9,13 @@ namespace Tests.FeatureManagement
         public const string TargetingTestFeatureWithExclusion = "TargetingTestFeatureWithExclusion";
         public const string OnTestFeature = "OnTestFeature";
         public const string OffTestFeature = "OffTestFeature";
-        public const string AlwaysOnTestFeature = "AlwaysOnTestFeature";
-        public const string OffTimeTestFeature = "OffTimeTestFeature";
         public const string ConditionalFeature = "ConditionalFeature";
-        public const string ConditionalFeature2 = "ConditionalFeature2";
         public const string ContextualFeature = "ContextualFeature";
         public const string AnyFilterFeature = "AnyFilterFeature";
         public const string AllFilterFeature = "AllFilterFeature";
         public const string FeatureUsesFiltersWithDuplicatedAlias = "FeatureUsesFiltersWithDuplicatedAlias";
+        public const string VariantFeatureDefaultDisabled = "VariantFeatureDefaultDisabled";
         public const string VariantFeatureDefaultEnabled = "VariantFeatureDefaultEnabled";
-        public const string VariantFeatureStatusDisabled = "VariantFeatureStatusDisabled";
         public const string VariantFeaturePercentileOn = "VariantFeaturePercentileOn";
         public const string VariantFeaturePercentileOff = "VariantFeaturePercentileOff";
         public const string VariantFeatureAlwaysOff = "VariantFeatureAlwaysOff";
@@ -31,5 +28,7 @@ namespace Tests.FeatureManagement
         public const string VariantFeatureInvalidStatusOverride = "VariantFeatureInvalidStatusOverride";
         public const string VariantFeatureInvalidFromTo = "VariantFeatureInvalidFromTo";
         public const string VariantImplementationFeature = "VariantImplementationFeature";
+        public const string OnTelemetryTestFeature = "OnTelemetryTestFeature";
+        public const string OffTelemtryTestFeature = "OffTelemetryTestFeature";
     }
 }
