+++ tests/Tests.FeatureManagement/Tests.FeatureManagement.csproj
@@ -46,7 +46,7 @@
     <None Update="appsettings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-    <None Update="MicrosoftFeatureManagement.json">
+    <None Update="DotnetFeatureManagementSchema.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
